requirements.txt
addict==2.4.0
alabaster==0.7.11
anaconda-client==1.7.2
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
batchgenerators==0.25
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==2.1.4
bokeh==0.13.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.3
cloudpickle==0.5.5
clyent==1.2.2
colorama==0.3.9
conda==4.5.11
conda-build==3.15.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.5
cytoolz==0.9.0.1
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==0.19.1
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
dicom2nifti==2.4.11
distributed==1.23.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fast_ctc_decode==0.3.5
fastcache==1.0.2
fastjsonschema==2.15.3
filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
future==1.0.0
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
imageio==2.4.1
imagesize==1.1.0
imgaug==0.4.0
incremental==17.5.0
ipykernel==4.9.0
ipython==6.5.0
ipython_genutils==0.2.0
ipywidgets==7.4.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.1
jeepney==0.3.1
Jinja2==2.10
jsonschema==2.6.0
jstyleson==0.0.2
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.34.9
jupyterlab-launcher==0.13.1
keyring==13.2.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
linecache2==1.0.0
llvmlite==0.24.0
lmdb==1.5.1
locket==0.2.0
loguru==0.7.2
lxml==4.2.5
MarkupSafe==2.1.2
matplotlib==2.2.3
mccabe==0.6.1
MedPy==0.4.0
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.6.3
nibabel==4.0.2
nltk==3.3
-e git+https://github.com/MIC-DKFZ/nnUNet.git@96d44c2fc1ce5e18da4cd54bf52882047c37982e#egg=nnunet
nose==1.3.7
notebook==5.6.0
numba==0.39.0
numexpr==2.6.8
numpy==1.19.5
numpydoc==0.8.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
odo==0.5.1
olefile==0.46
openai==0.8.0
opencv-python==4.10.0.84
openpyxl==2.5.6
openvino==2022.3.0
openvino-telemetry==2024.1.0
packaging==24.0
pandas==1.1.5
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
path.py==11.1.0
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==9.5.0
pkginfo==1.4.2
platformdirs==4.0.0
plotly==5.14.1
pluggy==0.7.1
ply==3.11
progress==1.6
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
py-cpuinfo==9.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyclipper==1.3.0.post5
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pydicom==2.4.4
PyExecJS==1.5.1
pyflakes==2.0.0
Pygments==2.2.0
pylint==2.1.1
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.8.0
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
python-gdcm==3.0.24.1
pytz==2024.2
PyWavelets==1.3.0
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.1
QtPy==1.5.0
regex==2024.4.16
requests==2.31.0
rope==0.11.0
ruamel_yaml==0.15.46
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==0.19.2
scipy==1.7.3
seaborn==0.9.0
SecretStorage==3.1.0
Send2Trash==1.5.0
sentencepiece==0.2.0
service-identity==17.0.0
shapely==2.0.5
simplegeneric==0.8.1
SimpleITK==2.3.0
singledispatch==3.4.0.3
singledispatchmethod==1.0
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.7.9
sphinxcontrib-websupport==1.1.0
spyder==3.3.1
spyder-kernels==0.2.6
SQLAlchemy==1.2.11
statsmodels==0.9.0
sympy==1.2
tables==3.4.4
tblib==1.3.2
tenacity==8.2.2
terminado==0.8.1
testpath==0.3.1
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.10.3
tomli==2.0.1
toolz==0.9.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==5.1
tqdm==4.66.4
traceback2==1.4.0
traitlets==4.3.2
Twisted==18.7.0
typing_extensions==4.7.1
unicodecsv==0.14.1
unittest2==1.1.0
urllib3==1.23
vtk==9.3.1
wcwidth==0.2.13
web-annotation-tool==0.0.1
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wrapt==1.10.11
xgboost==1.6.2
xlrd==1.1.0
XlsxWriter==1.1.0
xlwt==1.3.0
yapf==0.43.0
zict==0.1.3
zipp==3.15.0
zope.interface==4.5.0
